三国2战斗重置 - 战斗逻辑编写

+++ mw_fight/src/main/java/com/gryphpoem/game/zw/buff/impl/effect/ShieldEffectImpl.java
@@ -67,7 +67,7 @@ public class ShieldEffectImpl extends AbsFightEffect {
     }
 
     private int calEffectValue(Force force, int heroId, List<Integer> effectConfig) {
-        int attributeValue = FightCalc.attributeValue(effectConfig.get(3), force, heroId);
+        double attributeValue = FightCalc.attributeValue(effectConfig.get(3), force, heroId);
         return (int) (attributeValue * (effectConfig.get(4) / FightConstant.TEN_THOUSAND) + effectConfig.get(5));
     }
